% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_elements.R
\name{browse_elements}
\alias{browse_elements}
\title{Browse elements for description}
\usage{
browse_elements(pattern)
}
\arguments{
\item{pattern}{A perl expression or expressions to match in the long name as in \code{\link{heims_data_dict}}.}
}
\description{
Browse elements for description
}
