% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_using.R
\name{decode_using}
\alias{decode_using}
\title{Decode using data table}
\usage{
decode_using(DT, using)
}
\arguments{
\item{DT}{A \code{data.table} to decode.}

\item{using}{A keyed \code{data.table} with two columns. The key needs to be present in \code{DT}. The other column is column to remain.}
}
\description{
Decode using data table
}
