% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_enrol.R
\docType{data}
\name{dummy_enrol}
\alias{dummy_enrol}
\title{Dummy enrolment file}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 5 rows and 56 columns.
}
\usage{
dummy_enrol
}
\description{
A \code{data.table} of five fictitious enrolments.
}
\keyword{datasets}
