% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_heims.R
\name{element_decoders}
\alias{element_decoders}
\alias{rename_heims}
\alias{element2name}
\title{Make HEIMS element nos human-readable}
\usage{
rename_heims(DT)

element2name(v)
}
\arguments{
\item{DT}{The data table with original names}

\item{v}{A vector of element names.}
}
\value{
DT with the new names or the vector with the names translated. For \code{browse_elements}, a \code{data.table} of all element-long name combinations matching the perl regular expression.
}
\description{
Make HEIMS element nos human-readable
}
\details{
See \code{\link{heims_data_dict}}.
}
