% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_levels.R
\docType{data}
\name{first_levels}
\alias{first_levels}
\title{First levels}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 8 rows and 2 columns.
}
\usage{
first_levels
}
\description{
See \code{\link{relevel_heims}}.
}
\keyword{datasets}
