% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fread_heims.R
\name{fread_heims}
\alias{fread_heims}
\title{Read raw heims file}
\usage{
fread_heims(filename)
}
\arguments{
\item{filename}{A text-delimeted file, passed to \code{fread} from \code{data.table}.}
}
\value{
A \code{data.table}.
}
\description{
Read raw heims file
}
