% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_heims_fst.R
\name{read_heims_fst}
\alias{read_heims_fst}
\title{Read HEIMS data from decoded fst files}
\usage{
read_heims_fst(filename)
}
\arguments{
\item{filename}{File path to \code{.fst} file of a decoded HEIMS file (\code{\link{decode_heims}})
produced by \code{fst::write.fst}.}
}
\value{
A \code{data.table} with appropriate attributes.
}
\description{
Read HEIMS data from decoded fst files
}
