% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel_heims.R
\name{relevel_heims}
\alias{relevel_heims}
\title{Relevel categorical variables}
\usage{
relevel_heims(DT)
}
\arguments{
\item{DT}{A \code{data.table} post \code{\link{decode_heims}}.}
}
\value{
The same data.table with character vectors changed to factors whose first level is
the level intended.
}
\description{
Changes categorical variables in a data.table to levels with a sensible reference level
}
