% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_heims.R
\name{rename_heims}
\alias{rename_heims}
\title{Make HEIMS element nos human-readable}
\usage{
rename_heims(DT)
}
\arguments{
\item{DT}{The data table with original names}
}
\value{
DT with the new names
}
\description{
Make HEIMS element nos human-readable
}
\details{
See \code{\link{heims_data_dict}}.
}
