% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_heims.R
\name{decode_heims}
\alias{decode_heims}
\title{Decode HEIMS elements}
\usage{
decode_heims(DT, show_progress = FALSE)
}
\arguments{
\item{DT}{A \code{data.table} with the original column names.}

\item{show_progress}{Display the progress of the function (which is likely to be slow).}
}
\value{
DT with the values decoded and the names renamed.
}
\description{
Decode HEIMS elements
}
