% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-decoders.R
\docType{data}
\name{decoders}
\alias{decoders}
\alias{E490_decoder}
\title{Decoders}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 46 rows and 2 columns.}
\usage{
E490_decoder
}
\description{
Decoders
}
\keyword{datasets}
