% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_heims.R
\name{element_decoders}
\alias{element_decoders}
\alias{rename_heims}
\alias{element2name}
\title{Make HEIMS element nos human-readable}
\usage{
rename_heims(DT)

element2name(v)
}
\arguments{
\item{DT}{The data table with original names}

\item{v}{A vector of element names.}
}
\value{
\code{DT} with the new names or the vector with the names translated.
}
\description{
Make HEIMS element nos human-readable
}
\details{
See \code{\link{heims_data_dict}}. Note that \code{\link{decode_heims}} is generally better,
as it decodes the variable if a decoder is present in the dictionary.

\code{element2name} is the inverse of \code{\link{browse_elements}}:
given an element like \code{E306}, it returns
the name (\code{HE_Provider_cd}.)
}
