% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utilities}
\alias{utilities}
\alias{AND}
\alias{OR}
\alias{never}
\alias{every}
\alias{always}
\alias{is.Date}
\alias{is.YearMonth}
\alias{nth_digit_of}
\alias{between}
\alias{or}
\alias{and}
\alias{if_else}
\alias{coalesce}
\alias{\%fin\%}
\alias{rm_leading_0s}
\alias{as.integer64}
\alias{is.integer64}
\alias{force_integer}
\alias{ymd}
\title{Utility functions}
\usage{
AND()

OR()

never(v)

every(v)

always(v)

is.Date(v)

is.YearMonth(v)

nth_digit_of(x, n)

between(...)

or(...)

and(...)

if_else(...)

coalesce(...)

a \%fin\% tbl

rm_leading_0s(v)

as.integer64(v)

is.integer64(v)

force_integer(v)

ymd(...)
}
\arguments{
\item{v}{A vector.}

\item{x, n}{vectors}

\item{...}{Passed to other functions}

\item{a}{Element suspected to be in \code{tbl}}

\item{tbl}{A lookup table.}
}
\description{
Only included here because of the unusual nature of \code{\link{heims_data_dict}}.
}
\details{
\code{nth_digit_of} returns the nth digit of the number \strong{starting from the units and going up in magnitude.}
}
\examples{
nth_digit_of(503, 1) == 1
}
